+++ db/repos/index.js
@@ -17,23 +17,23 @@
  */
 
 module.exports = {
-	accounts: require('./accounts.js'),
-	blocks: require('./blocks.js'),
-	dapps: require('./dapps.js'),
-	delegates: require('./delegates.js'),
-	migrations: require('./migrations.js'),
-	multisignatures: require('./multisignatures.js'),
-	peers: require('./peers.js'),
-	rounds: require('./rounds.js'),
-	voters: require('./voters.js'),
-	votes: require('./votes.js'),
-	'transactions.dapp': require('./transactions.dapp.js'),
-	'transactions.delegate': require('./transactions.delegate.js'),
-	'transactions.inTransfer': require('./transactions.in_transfer.js'),
-	transactions: require('./transactions.js'),
-	'transactions.multisignature': require('./transactions.multisignature.js'),
-	'transactions.outTransfer': require('./transactions.out_transfer.js'),
-	'transactions.signature': require('./transactions.signature.js'),
-	'transactions.transfer': require('./transactions.transfer.js'),
-	'transactions.vote': require('./transactions.vote.js'),
+	accounts: require('./accounts'),
+	blocks: require('./blocks'),
+	dapps: require('./dapps'),
+	delegates: require('./delegates'),
+	migrations: require('./migrations'),
+	multisignatures: require('./multisignatures'),
+	peers: require('./peers'),
+	rounds: require('./rounds'),
+	voters: require('./voters'),
+	votes: require('./votes'),
+	transactions: require('./transactions'),
+	'transactions.dapp': require('./transactions/dapp'),
+	'transactions.delegate': require('./transactions/delegate'),
+	'transactions.inTransfer': require('./transactions/in_transfer'),
+	'transactions.multisignature': require('./transactions/multisignature'),
+	'transactions.outTransfer': require('./transactions/out_transfer'),
+	'transactions.signature': require('./transactions/signature'),
+	'transactions.transfer': require('./transactions/transfer'),
+	'transactions.vote': require('./transactions/vote'),
 };
